app/requirements.txt
boto3
requests
kaggle
pandas
mysql-connector-python
geopy
apache-airflow